+++ babel.config.js
@@ -4,7 +4,8 @@ module.exports = {
   plugins: [
     '@babel/plugin-syntax-jsx',
     '@babel/plugin-transform-flow-strip-types',
-    ['@babel/plugin-proposal-class-properties', {loose: true}],
+    ['@babel/plugin-transform-class-properties', {loose: true}],
+    '@babel/plugin-transform-classes',
     'syntax-trailing-function-commas',
     [
       '@babel/plugin-proposal-object-rest-spread',