+++ packages/shared/forks/ReactFeatureFlags.www.js
@@ -37,7 +37,9 @@ export const {
   transitionLaneExpirationMs,
   enableFastAddPropertiesInDiffing,
   enableViewTransition,
+  enableComponentPerformanceTrack,
   enableScrollEndPolyfill,
+  enableFragmentRefs,
 } = dynamicFeatureFlags;
 
 // On WWW, __EXPERIMENTAL__ is used for a new modern build.
@@ -62,8 +64,6 @@ export const enableYieldingBeforePassive = false;
 export const enableThrottledScheduling = false;
 
 export const enableHydrationLaneScheduling = true;
-
-export const enableComponentPerformanceTrack = false;
 
 // Logs additional User Timing API marks for use with an experimental profiling tool.
 export const enableSchedulingProfiler: boolean =